shared: qssquery: Fix sqlerror if sort condition comes last
If we had something like c:(test) sort:(mtime, desc), then
we would run into an sql error eventually, since we
would have added an implicit AND token before the sort
condition. This is wrong, as a sort is not a filter.

So, as a quick hack to fix this, just remove a preceeding
boolean (as dirty as it may be, but it does the job for now).